requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
duckdb>=0.10.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
apscheduler>=3.10.4
pyjwt>=2.8.0
bcrypt>=4.1.2
python-multipart>=0.0.9
pyyaml>=6.0.1
python-dotenv>=1.0.1
websockets>=12.0
